bors: initial MergeQueue logic
Implement the initial MergeQueue logic and add support for merging a PR
when its been approved and asked to be landed. This initial
implementation merges PRs via Rebasing them on top of master and then
updates the PR in-place with the newly rebased commits prior to pushing
them to master in order to have Github's UI correctly mark the PR as
being 'merged'.